pyproject.toml
[tool.black]
target-version = ["py312"]

[tool.isort]
profile = "black"
skip_glob = "*/migrations/*"
known_django = "django"
known_fastapi = "fastapi"
known_local_folder = "agents,backend,config"
sections = "FUTURE,STDLIB,FIRSTPARTY,THIRDPARTY,DJANGO,FASTAPI,LOCALFOLDER"

[tool.autoflake]
in-place = true
quiet = true
recursive = true
remove-all-unused-imports = true

[tool.ruff]
target-version = "py312"

[tool.poetry]
name = "cph24"
version = "0.1.0"
description = "Package supporting talk."
authors = [
    "Joseph V. Zammit <jvzammit@gmail.com>",
]
readme = "README.md"
[tool.poetry.dependencies]
python = "~3.12"
django = "^5.1.1"
arrow = "^1.3.0"
pydantic = "^2.9.2"
django-environ = "^0.11.2"
pydantic-settings = "^2.5.2"
fastapi = "^0.115.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.7"
factory-boy = "^3.3.1"
pre-commit = "^3.8.0"
black = "^24.8.0"
isort = "^5.13.2"
autoflake = "^2.3.1"
coverage = "^7.6.1"
mypy = "^1.11.2"
nox = "^2024.4.15"
fastapi-cli = "^0.0.5"
pytest-django = "^4.9.0"
django-extensions = "^3.2.3"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
warn_unused_ignores = true
warn_unreachable = true
plugins = [
    "pydantic.mypy"
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "*.tests.*"
]
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = [
    "*.migrations.*",
]
ignore_errors = true
